swaggy float array

+++ src/tests/t17.c
@@ -1,9 +1,9 @@
 float f(float x) {
-  int a;
+  float a;
   float b;
-  a=1000;
-  b=20123*x;
-  return a+b*3.-8;
+  a=1000.;
+  b=20123.*x;
+  return a+b*3.-8.;
 }
 int main() {
   float a;
